requirements.txt
pytube
moviepy
torch
openai-whisper
transformers
gradio
groq